requirements.txt
Django>=1.6.2
django-bootstrap3-datetimepicker>=2.2.3
django-braces>=1.3.1
django-crispy-forms>=1.4.0
pytz
six>=1.5.2
wsgiref>=0.1.2
django-floppyforms>=1.1.1
MySQL-python>=1.2.5
django-report-builder>=1.14